fix PanelViewPreviewConsequence.ts
