+++ app/frontend/src/home.ts
@@ -7,14 +7,14 @@ import ModuleTabsView from './components/ModuleTabsView';
 import DownloadButton from './components/DownloadButton';
 import SimpleSearchView from './components/SearchField/SimpleSearch/SimpleSearchView';
 import PanelViewCheckList from './components/PanelView/PanelViewCheckList';
-import PanelViewFilterAlternateAlleleFrequency from './components/PanelView/PanelViewFilterAlternateAlleleFrequency.js';
-import PanelViewFilterVariantCallingQuality from './components/PanelView/PanelViewFilterVariantCallingQuality.js';
+import PanelViewFilterAlternateAlleleFrequency from './components/PanelView/PanelViewFilterAlternateAlleleFrequency';
+import PanelViewFilterVariantCallingQuality from './components/PanelView/PanelViewFilterVariantCallingQuality';
 import PanelViewFilterConsequence from './components/PanelView/PanelViewFilterConsequence';
 import PanelViewPreviewGene from './components/PanelView/PanelViewPreviewGene.js';
 import PreviewToVariantReport from './components/PanelView/PreviewToVariantReport.js';
 import PanelViewPreviewExternalLinks from './components/PanelView/PanelViewPreviewExternalLinks.js';
-import PanelViewPreviewAlternateAlleleFrequencies from './components/PanelView/PanelViewPreviewAlternateAlleleFrequencies.js';
-import PanelViewPreviewConsequence from './components/PanelView/PanelViewPreviewConsequence.js';
+import PanelViewPreviewAlternateAlleleFrequencies from './components/PanelView/PanelViewPreviewAlternateAlleleFrequencies';
+import PanelViewPreviewConsequence from './components/PanelView/PanelViewPreviewConsequence';
 import PanelViewPreviewClinicalSignificance from './components/PanelView/PanelViewPreviewClinicalSignificance';
 import FloatingInfo from './components/FloatingInfo';
 import qs from 'qs';
